docs: tooltip demo and docs update
